refactor: refactored navbar service module
