+++ lib/features/interactive-book/services/navbar.dart
